# apps/backend/core-api/docs/swagger.yaml
definitions:
...
  certificate_share_handler.CertificateShareViewModel:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      event_certificate_id:
        type: string
      handle:
        type: string
      has_password:
        type: boolean
      id:
        type: string
      updated_at:
        type: string
    required:
    - active
    - created_at
    - event_certificate_id
    - handle
    - has_password
    - id
    - updated_at
    type: object
  certificate_share_handler.CreateCertificateShareBody:
    properties:
      password:
        type: string
    type: object
  certificate_share_handler.CreateCertificateShareResponse:
    properties:
      share:
        $ref: '#/definitions/certificate_share_handler.CertificateShareViewModel'
    required:
    - share
    type: object
...
  certificate_share_handler.GetCertificateShareImageBody:
    properties:
      password:
        type: string
    type: object
  certificate_share_handler.UpdateCertificateShareBody:
    properties:
      active:
        type: boolean
      password:
        type: string
    type: object
  certificate_share_handler.UpdateCertificateShareResponse:
    properties:
      share:
        $ref: '#/definitions/certificate_share_handler.CertificateShareViewModel'
    required:
    - share
    type: object
...
  core-api_internal_handler_event.EventResponse:
    properties:
      attendees_count:
        type: integer
      banner_storage_key:
        type: string
      chain_id:
        type: integer
      contact_address:
        type: string
      contact_number:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      event_status:
        $ref: '#/definitions/entity.EventStatus'
      event_type:
        $ref: '#/definitions/entity.EventType'
      google_map_query:
        type: string
      icon_storage_key:
        type: string
      id:
        type: string
      is_booking_request_required:
        type: boolean
      is_public:
        type: boolean
      is_ticket_transferable:
        type: boolean
      is_verified:
        type: boolean
      location:
        type: string
      long_description:
        type: string
      max_attendees:
        type: integer
      owner_credential_id:
        type: string
      short_description:
        type: string
      start_date:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - attendees_count
    - banner_storage_key
    - chain_id
    - contact_address
    - contact_number
    - created_at
    - end_date
    - event_status
    - event_type
    - google_map_query
    - icon_storage_key
    - id
    - is_booking_request_required
    - is_public
    - is_ticket_transferable
    - is_verified
    - location
    - long_description
    - max_attendees
    - owner_credential_id
    - short_description
    - start_date
    - title
    - updated_at
    type: object
...
  customerror.ErrResponse:
    description: Response for the client to sign to register
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  entity.Event:
    properties:
      attendees_count:
        type: integer
      banner_storage_key:
        type: string
      chain_id:
        type: integer
      contact_address:
        type: string
      contact_number:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      event_status:
        $ref: '#/definitions/entity.EventStatus'
      event_type:
        $ref: '#/definitions/entity.EventType'
      google_map_query:
        type: string
      icon_storage_key:
        type: string
      id:
        type: string
      is_booking_request_required:
        type: boolean
      is_public:
        type: boolean
      is_ticket_transferable:
        type: boolean
      is_verified:
        type: boolean
      location:
        type: string
      long_description:
        type: string
      max_attendees:
        type: integer
      owner_credential_id:
        type: string
      short_description:
        type: string
      start_date:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - attendees_count
    - banner_storage_key
    - chain_id
    - contact_address
    - contact_number
    - created_at
    - end_date
    - event_status
    - event_type
    - google_map_query
    - icon_storage_key
    - id
    - is_booking_request_required
    - is_public
    - is_ticket_transferable
    - is_verified
    - location
    - long_description
    - max_attendees
    - owner_credential_id
    - short_description
    - start_date
    - title
    - updated_at
    type: object
...
  entity.EventStatus:
    enum:
    - active
    - inactive
    - closed
    type: string
    x-enum-varnames:
    - EventStatusActive
    - EventStatusInactive
    - EventStatusClosed
  entity.EventType:
    enum:
    - private
    - invite
    type: string
    x-enum-varnames:
    - EventTypePrivate
    - EventTypeInvite
...
  event.GetEventListResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/entity.Event'
        type: array
    required:
    - events
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: DECM (Decentralized Event Management) platform API for NFT ticketing,
    digital credentials, and academic identity verification.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DECM Core
  version: "1.0"
paths:
...
  /api/v1/certificate-shares/{handle}/image:
    post:
      consumes:
      - application/json
      description: Returns a PNG certificate image for a share link. For password-protected
        shares, include the password in the request body.
      operationId: get-certificate-share-image
      parameters:
      - description: Share handle
        in: path
        name: handle
        required: true
        type: string
      - description: Optional password for password-protected shares
        in: body
        name: body
        schema:
          $ref: '#/definitions/certificate_share_handler.GetCertificateShareImageBody'
      produces:
      - image/png
      responses:
        "200":
          description: PNG certificate image
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Get certificate share image
      tags:
      - CertificateShares
  /api/v1/certificate-shares/config/{certificate_id}:
    post:
      consumes:
      - application/json
      description: Create a shareable link for a certificate. Only the certificate
        owner may call this endpoint. An optional password can be set to restrict
        access.
      operationId: create-certificate-share
      parameters:
      - description: Certificate ID
        in: path
        name: certificate_id
        required: true
        type: string
      - description: Optional password configuration
        in: body
        name: body
        schema:
          $ref: '#/definitions/certificate_share_handler.CreateCertificateShareBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/certificate_share_handler.CreateCertificateShareResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      security:
      - BearerAuth: []
      summary: Create certificate share link
      tags:
      - CertificateShares
  /api/v1/certificate-shares/config/{share_id}:
    patch:
      consumes:
      - application/json
      description: Update the password of an existing share link. Only the certificate
        owner may call this endpoint. Omit the password field to leave existing protection
        unchanged; set to empty string or null to remove protection; set to a non-empty
        string to set a new password.
      operationId: update-certificate-share
      parameters:
      - description: Share ID
        in: path
        name: share_id
        required: true
        type: string
      - description: Optional password configuration
        in: body
        name: body
        schema:
          $ref: '#/definitions/certificate_share_handler.UpdateCertificateShareBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/certificate_share_handler.UpdateCertificateShareResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      security:
      - BearerAuth: []
      summary: Update certificate share link
      tags:
      - CertificateShares
  /api/v1/certificates/{certificate_id}/image:
    get:
      description: Generates a PNG certificate image for the authenticated user's
        certificate
      operationId: generate-certificate-image
      parameters:
      - description: Certificate ID
        format: uuid
        in: path
        name: certificate_id
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: PNG certificate image
          schema:
            type: file
        "400":
          description: Invalid certificate ID
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "403":
          description: User does not own this certificate
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Certificate not found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      security:
      - CookieAuth: []
      summary: Generate certificate image for participant
      tags:
      - certificates
...
  /api/v1/events:
    get:
      description: Get events list
      operationId: get-events-list
      parameters:
      - default: true
        description: Include active events
        in: query
        name: include_active_events
        type: boolean
      - default: true
        description: Include inactive events
        in: query
        name: include_inactive_events
        type: boolean
      - default: false
        description: Include closed events
        in: query
        name: include_closed_events
        type: boolean
      - default: false
        description: Only user joined events
        in: query
        name: only_user_joined_events
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.GetEventListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Get events list
      tags:
      - Event
    post:
      consumes:
      - multipart/form-data
      description: Create a new event with banner image upload. Chain ID is automatically
        set from environment configuration.
      operationId: create-event
      parameters:
      - description: Event name
        in: formData
        name: name
        required: true
        type: string
      - description: Event short description
        in: formData
        name: short_description
        required: true
        type: string
      - description: Event description
        in: formData
        name: description
        required: true
        type: string
      - description: Start date (RFC3339 format)
        in: formData
        name: start_date
        required: true
        type: string
      - description: End date (RFC3339 format)
        in: formData
        name: end_date
        required: true
        type: string
      - description: Number of seats
        in: formData
        name: seats_count
        required: true
        type: integer
      - description: Contact number
        in: formData
        name: contact_number
        required: true
        type: string
      - description: Contact address
        in: formData
        name: contact_address
        required: true
        type: string
      - description: Location
        in: formData
        name: location
        required: true
        type: string
      - description: Google map query
        in: formData
        name: google_map_query
        required: true
        type: string
      - description: Event banner image (JPEG, PNG, WebP, max 10MB)
        in: formData
        name: banner
        required: true
        type: file
      - description: Event icon image (JPEG, PNG, WebP, max 10MB)
        in: formData
        name: icon
        required: true
        type: file
      - description: Host password
        in: formData
        name: host_password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/core-api_internal_handler_event.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Create a new event
      tags:
      - Event
...
swagger: "2.0"